refactor: move i18n files from no-meteor to node-app
