+++ imports/node-app/core-services/i18n/i18n/es.json
